hal: Fix APP Type mixer control value set fail
fix missmatch with int and long, on 64 bit arch,
int is 4 bytes, while long is 8 bytes, this will
cause memcpy copy wrong data. replace int with
ssize_t

+++ hal/audio_extn/utils.c
@@ -1008,7 +1008,7 @@ static int set_stream_app_type_mixer_ctrl(struct audio_device *adev,
 
     char mixer_ctl_name[MAX_LENGTH_MIXER_CONTROL_IN_INT];
     struct mixer_ctl *ctl;
-    int app_type_cfg[MAX_LENGTH_MIXER_CONTROL_IN_INT], len = 0, rc = 0;
+    ssize_t app_type_cfg[MAX_LENGTH_MIXER_CONTROL_IN_INT], len = 0, rc = 0;
     int snd_device_be_idx = -1;
 
     if (stream_type == PCM_PLAYBACK) {
